moving modules from mathics.algorithm to mathics.eval
