+++ mathics/eval/numbers/algebra/__init__.py
@@ -1,0 +1,4 @@
+# -*- coding: utf-8 -*-
+"""
+Implementation of  mathics.builtin.numbers.algebra
+"""
