+++ mathics/eval/numbers/algebra/simplify.py
